# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Libreria",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "description": "Retrieve a list of all books, including their authors and genres, cached for 5 minutes.",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "A list of books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Add a new book to the database.",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get book details",
                "description": "Retrieve the details of a specific book by ID, including its author and genre.",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "description": "Update the details of an existing book by ID.",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "description": "Soft delete a specific book by ID.",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
        "/api/books/{id}/picture": {
            "get": {
                "tags": [
                    "Books",
                    "Pictures"
                ],
                "summary": "Get a book's image",
                "description": "Retrieves the image associated with the specified book and returns it as a binary stream.",
                "operationId": "9f8f8ae20bf0fbdfb4b0074236bb923d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image retrieved successfully",
                        "content": {
                            "image/jpeg": {}
                        }
                    },
                    "404": {
                        "description": "Image not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books",
                    "Pictures"
                ],
                "summary": "Upload a book's image",
                "description": "Uploads an image for the specified book and updates its picture field with the image path.",
                "operationId": "17bd63d927b0bc581f57c8f7abca198e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "The image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Image uploaded successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookRequest": {
                "title": "BookRequest",
                "properties": {
                    "name": {
                        "description": "Name of the book",
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "author_id": {
                        "description": "ID of the author",
                        "type": "number",
                        "example": 1
                    },
                    "genre_id": {
                        "description": "ID of the genre",
                        "type": "number",
                        "example": 2
                    },
                    "year_publication": {
                        "description": "Year of publication",
                        "type": "number",
                        "example": 1925
                    },
                    "picture": {
                        "description": "URL of the book's picture or null",
                        "oneOf": [
                            {
                                "type": "string",
                                "example": "https://example.com/picture.jpg"
                            },
                            {
                                "type": "null",
                                "example": null
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "Book": {
                "title": "Book",
                "description": "Book model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "genre_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "year_publication": {
                        "type": "integer",
                        "example": 1925
                    },
                    "picture": {
                        "type": "string",
                        "example": "images/0b12345c-678d-90ef-a123-456789abcdef.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T01:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}